Refactor application forms
