+++ pkg/app/web/src/components/add-application-drawer.stories.tsx
@@ -1,13 +1,13 @@
 import React from "react";
-import { ApplicationFormDrawer } from "./application-form-drawer";
+import { AddApplicationDrawer } from "./add-application-drawer";
 import { action } from "@storybook/addon-actions";
 import { createDecoratorRedux } from "../../.storybook/redux-decorator";
 import { dummyEnv } from "../__fixtures__/dummy-environment";
 import { dummyPiped } from "../__fixtures__/dummy-piped";
 
 export default {
-  title: "APPLICATION/ApplicationFormDrawer",
-  component: ApplicationFormDrawer,
+  title: "APPLICATION/AddApplicationDrawer",
+  component: AddApplicationDrawer,
   decorators: [
     createDecoratorRedux({
       environments: {
@@ -27,21 +27,5 @@ export default {
 };
 
 export const overview: React.FC = () => (
-  <ApplicationFormDrawer
-    open
-    title="add application"
-    onSubmit={action("onSubmit")}
-    onClose={action("onClose")}
-    isProcessing={false}
-  />
-);
-
-export const isProcessing: React.FC = () => (
-  <ApplicationFormDrawer
-    open
-    title="add application"
-    onSubmit={action("onSubmit")}
-    onClose={action("onClose")}
-    isProcessing={true}
-  />
+  <AddApplicationDrawer open onClose={action("onClose")} />
 );
